+++ core/site.ts
@@ -134,6 +134,7 @@ export default class Site {
 
   constructor(options: Partial<SiteOptions> = {}) {
     this.options = merge(defaults, options);
+    this.options.location.port ||= this.options.server.port.toString();
 
     const src = this.src();
     const dest = this.dest();