+++ core/utils/cli_options.ts
@@ -2,41 +2,41 @@ import { parseArgs } from "../../deps/cli.ts";
 import type { DeepPartial } from "./object.ts";
 import type { SiteOptions } from "../site.ts";
 
-export function getOptionsFromCli(): DeepPartial<SiteOptions> {
-  const options = parseArgs(Deno.args, {
+export function getOptionsFromCli(
+  options: DeepPartial<SiteOptions>,
+): DeepPartial<SiteOptions> {
+  const cli = parseArgs(Deno.args, {
     string: ["src", "dest", "location", "port"],
     boolean: ["serve", "open"],
     alias: { dev: "d", serve: "s", port: "p", open: "o" },
     ["--"]: true,
   });
 
-  const overrides: DeepPartial<SiteOptions> = {};
-
-  if (options.src) {
-    overrides.src = options.src;
+  if (cli.src) {
+    options.src = cli.src;
   }
 
-  if (options.dest) {
-    overrides.dest = options.dest;
+  if (cli.dest) {
+    options.dest = cli.dest;
   }
 
-  if (options.location) {
-    overrides.location = new URL(options.location);
-  } else if (options.serve || options._[0] === "cms") {
-    overrides.location = new URL(`http://localhost:${options.port || 3000}/`);
+  if (cli.location) {
+    options.location = new URL(cli.location);
+  } else if (cli.serve || cli._[0] === "cms") {
+    options.location = new URL(`http://localhost:${cli.port || 3000}/`);
   }
 
-  if (options.port) {
-    (overrides.server ||= {}).port = parseInt(options.port);
+  if (cli.port) {
+    (options.server ||= {}).port = parseInt(cli.port);
 
-    if (overrides.location) {
-      overrides.location.port = options.port;
+    if (options.location) {
+      options.location.port = cli.port;
     }
   }
 
-  if (options.open) {
-    (overrides.server ||= {}).open = options.open;
+  if (cli.open) {
+    (options.server ||= {}).open = cli.open;
   }
 
-  return overrides;
+  return options;
 }